Add files via upload
In Turkish, one does not need to use the pronoun 'I' to speak about oneself. This has made its literature of love very different from English: a language of I-want, I-think and I-feel, English often centers love around a give-and-take between two clear parties - I and you. Turkish, on the other hand, nurtures a self-effacing love: "Let me be the wind around your slender waist," it begs, "the land of grass you tread." I wanted to hear this difference that, in reading, I can only feel; this program is made to sonify poetry, associating a long, high tonic note with the pronoun 'I'. All other words sound as random notes pulled from the specified key. The program returns a midi file that can be saved to disk. It works currently for English, Turkish, German and Spanish.